+++ components/browser/menu/src/main/java/mozilla/components/browser/menu/item/BrowserMenuCheckbox.kt
@@ -4,6 +4,12 @@
 
 package mozilla.components.browser.menu.item
 
+import android.view.View
+import android.widget.LinearLayout
+import android.widget.TextView
+import androidx.appcompat.widget.AppCompatCheckBox
+import mozilla.components.browser.menu.BrowserMenu
+import mozilla.components.browser.menu.BrowserMenuItem
 import mozilla.components.browser.menu.R
 
 /**
@@ -14,9 +20,36 @@ import mozilla.components.browser.menu.R
  * @param listener Callback to be invoked when this menu item is checked.
  */
 class BrowserMenuCheckbox(
-    label: String,
-    initialState: () -> Boolean = { false },
-    listener: (Boolean) -> Unit
-) : BrowserMenuCompoundButton(label, initialState, listener) {
+    private val label: String,
+    private val initialState: () -> Boolean = { false },
+    private val listener: (Boolean) -> Unit
+) : BrowserMenuItem {
+
     override fun getLayoutResource() = R.layout.mozac_browser_menu_item_checkbox
+
+    override val visible: () -> Boolean = { true }
+
+    override fun bind(menu: BrowserMenu, view: View) {
+        (view as LinearLayout).apply {
+
+            val textView = findViewById<TextView>(R.id.text)
+            val checkbox = findViewById<AppCompatCheckBox>(R.id.checkbox)
+            textView.text = label
+            checkbox.isChecked = initialState()
+
+            // We delay the dismissing of the menu so the user has feedback the checkbox is selected
+            // This is to maintain parity with how this feature worked prior to its refactor
+            setOnClickListener {
+                checkbox.performClick()
+                listener(checkbox.isChecked)
+                handler.postDelayed({
+                    menu.dismiss()
+                }, DISMISS_DELAY)
+            }
+        }
+    }
+
+    private companion object {
+        const val DISMISS_DELAY = 250L
+    }
 }